Minimize breaks swap: minimal config
Minimal config demonstrating the issue described at:
https://github.com/xmonad/xmonad-contrib/issues/374

I ran to following commands to create this commit:

  git clone --recursive git@github.com:ivanbrennan/xmonad-testing.git
  cd xmonad-testing
  git checkout -b minimize-breaks-swap
  git submodule update --remote --merge

  # add 'containers' to build-depends
  vim xmonad-testing.cabal

  # create test config
  vim configs/minimize-breaks-swap.hs
  ln -sf configs/minimize-breaks-swap.hs config.hs

  # create development environment
  vim shell.nix

  # fixup Xephyr script
  vim bin/run-in-xephyr.sh

  # build
  nix-shell --run "cabal new-build"

  git add x11 xmonad xmonad-contrib xmonad-testing.cabal shell.nix config.hs configs/minimize-breaks-swap.hs bin/run-in-xephyr.sh
  git commit

I then ran an isolated xmonad session with bin/run-in-xephyr.sh

Within that session, I opened 3 terminal windows and verified that
swapUp and swapDown behaved as expected in the absence of any minimized
windows.

I then navigated to the window in the middle of the stack, minimized it,
and proceeded to attempt to swap the remainin windows up and down. The
swap behavior exhibited interference from the minimized window. For
example, with the master window focused, a single swapDown had no
apparent effect. A second swapDown did have the desired effect, so it
seems the first swap performed some behind-the-scenes state change
involving the master window and the minimized window.